% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covreg_reduced.R
\name{covreg_reduced}
\alias{covreg_reduced}
\title{Reduced covariance regression model using Stan}
\usage{
covreg_reduced(data, response, condition, participant, seed = 112243)
}
\description{
Reduced covariance regression model using Stan
}
