% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coeff.R
\name{plot_coeff}
\alias{plot_coeff}
\title{Calculate parcel set and plot coefficient.}
\usage{
plot_coeff(fit, response, alpha = 0.05, title = "", brain_slices = NULL,
  FDR = FALSE, brain_nrow = 3, coeff_labels = response)
}
\description{
Calculate parcel set and plot coefficient.
}
