% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_explained_cov.R
\name{plot_explained_cov}
\alias{plot_explained_cov}
\title{Plot unexplained covariance part.}
\usage{
plot_explained_cov(fit, data, response, condition, channel_location, x_label,
  y_label)
}
\description{
Plot unexplained covariance part.
}

