% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_explained_cov_channels.R
\name{plot_explained_cov_channels}
\alias{plot_explained_cov_channels}
\title{Plot channel explained cov in brain coordinates.}
\usage{
plot_explained_cov_channels(fit, data, response, condition, channel_location,
  x_label, y_label)
}
\description{
Plot channel explained cov in brain coordinates.
}

