% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_post_mean.R
\name{plot_post_mean}
\alias{plot_post_mean}
\title{Plot posterior mean correlation and standard deviation.}
\usage{
plot_post_mean(fit, channel_names)
}
\description{
Plot posterior mean correlation and standard deviation.
}
