% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_unexplained_cov_diagnostics.R
\name{plot_unexplained_cov_diagnostics}
\alias{plot_unexplained_cov_diagnostics}
\title{Trace plots of unexplained covariance parameters.}
\usage{
plot_unexplained_cov_diagnostics(fit, num_par = 8)
}
\description{
Trace plots of unexplained covariance parameters.
}

