% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covreg_low.R
\name{covreg_low}
\alias{covreg_low}
\title{Low-dimensional covariance regression model using Stan.}
\usage{
covreg_low(data, response, condition, participant, seed = 112243)
}
\description{
Low-dimensional covariance regression model using Stan.
}
